Randomizing zip code selection, fixing placement
Previously, when a person was generated in a city with multiple zip
codes, Synthea would select the first zip code for the city.

This change causes it to randomly choose a zip code.

This also changes how coordinates are randomized. Previously, it would
always add to the lat and long. This had the effect of moving the
population north and east. This change adds and subtracts distance to
randomize around the desired point.